Add item names in checkout and new file for types
