+++ app/paymentprovider/vivawallet.go
@@ -9,86 +9,28 @@ import (
 	"errors"
 	"io"
 	"strconv"
+	"strings"
 
 	"net/http"
 	"net/url"
 	"time"
+
+	"go.uber.org/zap"
 )
 
 var log = utils.GetLogger()
 
-type PaymentOrderRequest struct {
-	Amount              int      `json:"amount"`
-	CustomerTrns        string   `json:"customerTrns"`
-	Customer            Customer `json:"customer"`
-	PaymentTimeout      int      `json:"paymentTimeout"`
-	Preauth             bool     `json:"preauth"`
-	AllowRecurring      bool     `json:"allowRecurring"`
-	MaxInstallments     int      `json:"maxInstallments"`
-	PaymentNotification bool     `json:"paymentNotification"`
-	TipAmount           int      `json:"tipAmount"`
-	DisableExactAmount  bool     `json:"disableExactAmount"`
-	DisableCash         bool     `json:"disableCash"`
-	DisableWallet       bool     `json:"disableWallet"`
-	SourceCode          string   `json:"sourceCode"`
-	MerchantTrns        string   `json:"merchantTrns"`
-	Tags                []string `json:"tags"`
-	// TODO: Additional fields that Aaron found in the API docs
-	// PaymentMethodFees   []struct {
-	// 	PaymentMethodId int `json:"paymentMethodId"`
-	// 	Fee             int `json:"fee"`
-	// } `json:"paymentMethodFees"`
-	// CardTokens []string `json:"cardTokens"`
-}
-
-type Customer struct {
-	Email       string `json:"email"`
-	Fullname    string `json:"fullName"`
-	Phone       string `json:"phone"`
-	CountryCode string `json:"countryCode"`
-	RequestLang string `json:"requestLang"`
-}
-
-type AuthenticationResponse struct {
-	AccessToken string `json:"access_token"`
-	ExpiresIn   int    `json:"expires_in"`
-	TokenType   string `json:"token_type"`
-	Scope       string `json:"scope"`
-}
-
-type PaymentOrderResponse struct {
-	OrderCode int `json:"orderCode"`
-}
-
-type TransactionVerificationResponse struct {
-	Email               string  `json:"email"`
-	Amount              float64 `json:"amount"`
-	OrderCode           int     `json:"orderCode"`
-	StatusId            string  `json:"statusId"`
-	FullName            string  `json:"fullName"`
-	InsDate             string  `json:"insDate"`
-	CardNumber          string  `json:"cardNumber"`
-	CurrencyCode        string  `json:"currencyCode"`
-	CustomerTrns        string  `json:"customerTrns"`
-	MerchantTrns        string  `json:"merchantTrns"`
-	TransactionTypeId   int     `json:"transactionTypeId"`
-	RecurringSupport    bool    `json:"recurringSupport"`
-	TotalInstallments   int     `json:"totalInstallments"`
-	CardCountryCode     string  `json:"cardCountryCode"`
-	CardIssuingBank     string  `json:"cardIssuingBank"`
-	CurrentInstallment  int     `json:"currentInstallment"`
-	CardUniqueReference string  `json:"cardUniqueReference"`
-	CardTypeId          int     `json:"cardTypeId"`
-}
-
 func AuthenticateToVivaWallet() (string, error) {
 	// Create a new request URL using http
-	apiURL := "https://demo-accounts.vivapayments.com"
+	apiURL := config.Config.VivaWalletAccountsURL
+	if apiURL == "" {
+		return "", errors.New("VivaWalletAccountURL is not set")
+	}
 	resource := "/connect/token"
 	jsonPost := []byte(`grant_type=client_credentials`)
 	u, _ := url.ParseRequestURI(apiURL)
 	u.Path = resource
-	urlStr := u.String() // "https://demo-accounts.vivapayments.com/connect/token"
+	urlStr := u.String()
 
 	req, err := http.NewRequest("POST", urlStr, bytes.NewReader(jsonPost))
 	if err != nil {
@@ -106,8 +48,8 @@ func AuthenticateToVivaWallet() (string, error) {
 
 	// Create a new client with a 10 second timeout
 	client := http.Client{Timeout: 10 * time.Second}
-	// Send the request
 
+	// Send the request
 	res, err := client.Do(req)
 	if err != nil {
 		log.Error("impossible to send request: ", err)
@@ -115,6 +57,7 @@ func AuthenticateToVivaWallet() (string, error) {
 
 	// Close the body after the function returns
 	defer res.Body.Close()
+
 	// Log the request body
 	body, err := io.ReadAll(res.Body)
 	if err != nil {
@@ -133,16 +76,19 @@ func AuthenticateToVivaWallet() (string, error) {
 	return authResponse.AccessToken, nil
 }
 
-func CreatePaymentOrder(accessToken string, amount int) (int, error) {
+func CreatePaymentOrder(accessToken string, order database.Order) (int, error) {
 	// Create a new request URL using http
-	apiURL := "https://demo-api.vivapayments.com"
+	apiURL := config.Config.VivaWalletApiURL
+	if apiURL == "" {
+		return 0, errors.New("VivaWalletApiURL is not set")
+	}
 	resource := "/checkout/v2/orders"
 	u, _ := url.ParseRequestURI(apiURL)
 	u.Path = resource
-	urlStr := u.String() // "https://demo-accounts.vivapayments.com/connect/token"
+	urlStr := u.String()
 
 	// Create a new sample customer
-	// TODO: Change this to a real customer
+	// TODO once registration is possible: Check if user is "UserAnon" and if not change this to customer fields
 	customer := Customer{
 		Email:       "verein@augustin.or.at",
 		Fullname:    "Augustin Straßenzeitung",
@@ -150,11 +96,22 @@ func CreatePaymentOrder(accessToken string, amount int) (int, error) {
 		RequestLang: "de-AT",
 	}
 
+	// Create string slice listing every item name in order
+	items := []string{}
+
+	// Iterate through the order entries and retrieve item names
+	for _, entry := range order.Entries {
+		item, err := database.Db.GetItem(entry.Item) // Replace this with your logic
+		if err != nil {
+			log.Error("Item could not be found", zap.Error(err))
+		}
+		items = append(items, item.Name)
+	}
+
 	// Create a new sample payment order
-	// TODO: Change this to a real payment order
 	paymentOrderRequest := PaymentOrderRequest{
-		Amount:              amount,
-		CustomerTrns:        "Augustin Straßenzeitung",
+		Amount:              order.GetTotal(),
+		CustomerTrns:        strings.Join(items, ", "),
 		Customer:            customer,
 		PaymentTimeout:      300,
 		Preauth:             false,
@@ -166,9 +123,8 @@ func CreatePaymentOrder(accessToken string, amount int) (int, error) {
 		DisableCash:         true,
 		DisableWallet:       true,
 		SourceCode:          utils.GetEnv("VIVA_WALLET_SOURCE_CODE", ""),
-		MerchantTrns:        "Die Augustin Familie bedankt sich für Ihre Spende!",
-		//TODO: Change tags to item name
-		Tags: []string{"augustin", "spende"},
+		MerchantTrns:        "Die Augustin Familie bedankt sich für Ihre Überweisung!",
+		Tags:                items,
 	}
 
 	// Create a new post request
@@ -218,121 +174,14 @@ func CreatePaymentOrder(accessToken string, amount int) (int, error) {
 
 }
 
-type PriceEventData struct {
-	OrderCode       int64   `json:"OrderCode"`
-	MerchantId      string  `json:"MerchantId"`
-	IsvFee          float64 `json:"IsvFee"`
-	TransactionId   string  `json:"TransactionId"`
-	CurrencyCode    string  `json:"CurrencyCode"`
-	Interchange     float64 `json:"Interchange"`
-	TotalCommission float64 `json:"TotalCommission"`
-}
-
-type TransactionPriceRequest struct {
-	Url           string         `json:"Url"`
-	EventData     PriceEventData `json:"EventData"`
-	Created       time.Time      `json:"Created"`
-	CorrelationId string         `json:"CorrelationId"`
-	EventTypeId   int            `json:"EventTypeId"`
-	Delay         *int           `json:"Delay"`
-	MessageId     string         `json:"MessageId"`
-	RecipientId   string         `json:"RecipientId"`
-	MessageTypeId int            `json:"MessageTypeId"`
-}
-
-type TransactionDetailRequest struct {
-	Url           string    `json:"Url"`
-	EventData     EventData `json:"EventData"`
-	Created       time.Time `json:"Created"`
-	CorrelationId string    `json:"CorrelationId"`
-	EventTypeId   int       `json:"EventTypeId"`
-	Delay         any       `json:"Delay"`
-	MessageId     string    `json:"MessageId"`
-	RecipientId   string    `json:"RecipientId"`
-	MessageTypeId int       `json:"MessageTypeId"`
-}
-
-type EventData struct {
-	Moto                        bool     `json:"Moto"`
-	BinId                       int      `json:"BinId"`
-	Ucaf                        string   `json:"Ucaf"`
-	Email                       string   `json:"Email"`
-	Phone                       string   `json:"Phone"`
-	BankId                      string   `json:"BankId"`
-	Systemic                    bool     `json:"Systemic"`
-	Switching                   bool     `json:"Switching"`
-	ParentId                    any      `json:"ParentId"`
-	Amount                      float64  `json:"Amount"`
-	ChannelId                   string   `json:"ChannelId"`
-	TerminalId                  int      `json:"TerminalId"`
-	MerchantId                  string   `json:"MerchantId"`
-	OrderCode                   int      `json:"OrderCode"`
-	ProductId                   any      `json:"ProductId"`
-	StatusId                    string   `json:"StatusId"`
-	FullName                    string   `json:"FullName"`
-	ResellerId                  any      `json:"ResellerId"`
-	DualMessage                 bool     `json:"DualMessage"`
-	InsDate                     string   `json:"InsDate"`
-	TotalFee                    float64  `json:"TotalFee"`
-	CardToken                   string   `json:"CardToken"`
-	CardNumber                  string   `json:"CardNumber"`
-	TipAmount                   float64  `json:"TipAmount"`
-	SourceCode                  string   `json:"SourceCode"`
-	SourceName                  string   `json:"SourceName"`
-	Latitude                    any      `json:"Latitude"`
-	Longitude                   any      `json:"Longitude"`
-	CompanyName                 any      `json:"CompanyName"`
-	TransactionId               string   `json:"TransactionId"`
-	CompanyTitle                any      `json:"CompanyTitle"`
-	PanEntryMode                string   `json:"PanEntryMode"`
-	ReferenceNumber             int      `json:"ReferenceNumber"`
-	ResponseCode                string   `json:"ResponseCode"`
-	CurrencyCode                string   `json:"CurrencyCode"`
-	OrderCulture                string   `json:"OrderCulture"`
-	MerchantTrns                string   `json:"MerchantTrns"`
-	CustomerTrns                string   `json:"CustomerTrns"`
-	IsManualRefund              bool     `json:"IsManualRefund"`
-	TargetPersonId              any      `json:"TargetPersonId"`
-	TargetWalletId              any      `json:"TargetWalletId"`
-	AcquirerApproved            bool     `json:"AcquirerApproved"`
-	LoyaltyTriggered            bool     `json:"LoyaltyTriggered"`
-	TransactionTypeId           int      `json:"TransactionTypeId"`
-	AuthorizationId             string   `json:"AuthorizationId"`
-	TotalInstallments           int      `json:"TotalInstallments"`
-	CardCountryCode             any      `json:"CardCountryCode"`
-	CardIssuingBank             any      `json:"CardIssuingBank"`
-	RedeemedAmount              float64  `json:"RedeemedAmount"`
-	ClearanceDate               any      `json:"ClearanceDate"`
-	CurrentInstallment          int      `json:"CurrentInstallment"`
-	Tags                        []string `json:"Tags"`
-	BillId                      any      `json:"BillId"`
-	ConnectedAccountId          any      `json:"ConnectedAccountId"`
-	ResellerSourceCode          any      `json:"ResellerSourceCode"`
-	ResellerSourceName          any      `json:"ResellerSourceName"`
-	MerchantCategoryCode        int      `json:"MerchantCategoryCode"`
-	ResellerCompanyName         any      `json:"ResellerCompanyName"`
-	CardUniqueReference         string   `json:"CardUniqueReference"`
-	ResellerSourceAddress       any      `json:"ResellerSourceAddress"`
-	CardExpirationDate          string   `json:"CardExpirationDate"`
-	ServiceId                   any      `json:"ServiceId"`
-	RetrievalReferenceNumber    string   `json:"RetrievalReferenceNumber"`
-	AssignedMerchantUsers       []any    `json:"AssignedMerchantUsers"`
-	AssignedResellerUsers       []any    `json:"AssignedResellerUsers"`
-	CardTypeId                  int      `json:"CardTypeId"`
-	ResponseEventId             any      `json:"ResponseEventId"`
-	ElectronicCommerceIndicator string   `json:"ElectronicCommerceIndicator"`
-	OrderServiceId              int      `json:"OrderServiceId"`
-	DigitalWalletId             any      `json:"DigitalWalletId"`
-}
-
 func HandlePaymentSuccessfulResponse(paymentSuccessful TransactionDetailRequest) (err error) {
 
 	// Set everything up for the request
 
 	// Create a new request URL using http
-	apiURL := config.Config.VivaWalletURL
+	apiURL := config.Config.VivaWalletApiURL
 	if apiURL == "" {
-		return errors.New("VivaWalletURL is not set")
+		return errors.New("VivaWalletApiURL is not set")
 	}
 	// Use transactionId from webhook to get transaction details
 	resource := "/checkout/v2/transactions/" + paymentSuccessful.EventData.TransactionId
